Drone_Frame used for Demo
We removed the motor holder parts to reduce the weight and super glued the motors to the frame. 
Also, gyro MPU 6050 was glued to the frame. Between Pi Pico and motor driver, we added a wood plate to prevent short.
Battery was placed at the bottom bed of this drone and cable tied.